@@ -33,7 +33,7 @@ import org.keycloak.testsuite.arquillian.annotation.RestartContainer;
 import org.keycloak.testsuite.auth.page.WelcomePage;
 import org.keycloak.testsuite.auth.page.login.OIDCLogin;
 import org.keycloak.testsuite.util.DroneUtils;
-import org.keycloak.testsuite.util.PhantomJSBrowser;
+import org.keycloak.testsuite.util.HtmlUnitBrowser;
 import org.openqa.selenium.WebDriver;
 
 import java.net.InetAddress;
@@ -54,15 +54,15 @@ import static org.keycloak.testsuite.util.URLUtils.navigateToUri;
 public class WelcomePageTest extends AbstractKeycloakTest {
 
     @Drone
-    @PhantomJSBrowser
-    private WebDriver phantomJS;
+    @HtmlUnitBrowser
+    private WebDriver htmlUnit;
 
     @Page
-    @PhantomJSBrowser
+    @HtmlUnitBrowser
     protected OIDCLogin loginPage;
 
     @Page
-    @PhantomJSBrowser
+    @HtmlUnitBrowser
     protected WelcomePage welcomePage;
 
     @Override
@@ -80,7 +80,7 @@ public class WelcomePageTest extends AbstractKeycloakTest {
         Assume.assumeThat("Test skipped",
                 suiteContext.getAuthServerInfo().isJBossBased(),
                 Matchers.is(true));
-        DroneUtils.replaceDefaultWebDriver(this, phantomJS);
+        DroneUtils.replaceDefaultWebDriver(this, htmlUnit);
         setDefaultPageUriParameters();
     }
 
@@ -150,8 +150,7 @@ public class WelcomePageTest extends AbstractKeycloakTest {
     public void test_5_AccessCreatedAdminAccount() throws Exception {
         welcomePage.navigateTo();
         welcomePage.navigateToAdminConsole();
-        // TODO PhantomJS is not loading the new admin console for some reason, so is not redirecting to the login page. It works with Chrome though.
-        Assert.assertEquals("Keycloak Administration Console", phantomJS.getTitle());
+        Assert.assertEquals("Keycloak Administration Console", htmlUnit.getTitle());
     }
 
     @Test
